[MM-37460] - Add tip for downloading desktop app to the Getting Started page
