+++ components/next_steps_view/next_steps_tips.tsx
@@ -17,10 +17,8 @@ import MoreChannels from 'components/more_channels';
 import TeamMembersModal from 'components/team_members_modal';
 import MarketplaceModal from 'components/plugin_marketplace';
 import RemoveNextStepsModal from 'components/sidebar/sidebar_next_steps/remove_next_steps_modal';
-import downloadApps from 'images/download-app.svg';
 
 import {browserHistory} from 'utils/browser_history';
-import * as UserAgent from 'utils/user_agent';
 import {
     ModalIdentifiers,
     RecommendedNextSteps,
@@ -33,60 +31,7 @@ import {getAnalyticsCategory} from './step_helpers';
 import IncidentsSvg from './images/incidents.svg';
 import DocumentsSvg from './images/documents.svg';
 import PluginsSvg from './images/plugins.svg';
-
-const seeAllApps = (isAdmin: boolean) => {
-    trackEvent(getAnalyticsCategory(isAdmin), 'cloud_see_all_apps');
-    window.open('https://mattermost.com/download/#mattermostApps', '_blank');
-};
-
-const downloadLatest = (isAdmin: boolean) => {
-    const baseLatestURL = 'https://latest.mattermost.com/mattermost-desktop-';
-
-    if (UserAgent.isWindows()) {
-        trackEvent(getAnalyticsCategory(isAdmin), 'click_download_app', {app: 'windows'});
-        window.open(`${baseLatestURL}exe`, '_blank');
-        return;
-    }
-
-    if (UserAgent.isMac()) {
-        trackEvent(getAnalyticsCategory(isAdmin), 'click_download_app', {app: 'mac'});
-        window.open(`${baseLatestURL}dmg`, '_blank');
-        return;
-    }
-
-    // TODO: isLinux?
-
-    seeAllApps(isAdmin);
-};
-
-const getDownloadButtonString = () => {
-    if (UserAgent.isWindows()) {
-        return (
-            <FormattedMessage
-                id='next_steps_view.tips.getForWindows'
-                defaultMessage='Get Mattermost for Windows'
-            />
-        );
-    }
-
-    if (UserAgent.isMac()) {
-        return (
-            <FormattedMessage
-                id='next_steps_view.tips.getForMac'
-                defaultMessage='Get Mattermost for Mac'
-            />
-        );
-    }
-
-    // TODO: isLinux?
-
-    return (
-        <FormattedMessage
-            id='next_steps_view.tips.getForDefault'
-            defaultMessage='Get Mattermost'
-        />
-    );
-};
+import DownloadSection from './download_section';
 
 const openAdminConsole = (isAdmin: boolean) => {
     trackEvent(getAnalyticsCategory(isAdmin), 'click_admin_console');
@@ -264,54 +209,6 @@ export default function NextStepsTips(props: Props) {
         );
     }
 
-    let downloadSection;
-    if (Utils.isMobile()) {
-        downloadSection = (
-            <div className='NextStepsView__tipsMobileMessage'>
-                <Card expanded={true}>
-                    <div className='Card__body'>
-                        <i className='icon icon-laptop'/>
-                        <FormattedMessage
-                            id='next_steps_view.mobile_tips_message'
-                            defaultMessage='To configure your workspace, continue on a desktop computer.'
-                        />
-                    </div>
-                </Card>
-            </div>
-        );
-    } else if (!UserAgent.isDesktopApp()) {
-        downloadSection = (
-            <div className='NextStepsView__download'>
-                <img src={downloadApps}/>
-                <div className='NextStepsView__downloadText'>
-                    <h4>
-                        <FormattedMessage
-                            id='next_steps_view.downloadDesktopAndMobile'
-                            defaultMessage='Download the Desktop and Mobile apps'
-                        />
-                    </h4>
-                    <div className='NextStepsView__downloadButtons'>
-                        <button
-                            className='NextStepsView__button NextStepsView__downloadForPlatformButton secondary'
-                            onClick={() => downloadLatest(props.isFirstAdmin)}
-                        >
-                            {getDownloadButtonString()}
-                        </button>
-                        <button
-                            className='NextStepsView__button NextStepsView__downloadAnyButton tertiary'
-                            onClick={() => seeAllApps(props.isFirstAdmin)}
-                        >
-                            <FormattedMessage
-                                id='next_steps_view.seeAllTheApps'
-                                defaultMessage='See all the apps'
-                            />
-                        </button>
-                    </div>
-                </div>
-            </div>
-        );
-    }
-
     let channelsSection;
     if (props.isFirstAdmin) {
         channelsSection = (
@@ -407,7 +304,7 @@ export default function NextStepsTips(props: Props) {
                     {nonMobileTips}
                     {channelsSection}
                 </div>
-                {downloadSection}
+                <DownloadSection isFirstAdmin={props.isFirstAdmin}/>
             </div>
         </div>
     );
