+++ components/next_steps_view/next_steps_view.tsx
@@ -20,6 +20,7 @@ import loadingIcon from 'images/spinner-48x48-blue.apng';
 import {StepType} from './steps';
 import './next_steps_view.scss';
 import NextStepsTips from './next_steps_tips';
+import DownloadSection from './download_section';
 import OnboardingBgSvg from './images/onboarding-bg-svg';
 import GettingStartedSvg from './images/getting-started-svg';
 import CloudLogoSvg from './images/cloud-logo-svg';
@@ -321,6 +322,7 @@ export default class NextStepsView extends React.PureComponent<Props, State> {
                         <GettingStartedSvg/>
                     </div>
                 </div>
+                <DownloadSection isFirstAdmin={this.props.isFirstAdmin}/>
             </div>
         );
     }
